Raise a clearer error if design review number is not set
If `post-dr-comment` is called in a run which is not on a design review,
the DR number is not set in the context. This previously led to an
incomprehensible error. This commit checks early to make sure the design
review number is set and errors with an explanation if it isn't.